requirements.txt
Django==3.1
celery==4.4.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
drf-yasg==1.17.1
django-silk==4.1.0
django-filter==2.4.0
psycopg2-binary==2.9.1 
Pillow==8.1.0
gunicorn==20.1.0
django-environ==0.7.0
django-split-settings==1.1.0
daphne==0.8.1
twisted==21.7.0